Move further unit tests
Issue-Id: #229
